Risk and Returns: The Sharpe Ratio
When we  assess whether to invest in an asset, we  want to look not only at how much money we could make but also at how much risk we are taking. The Sharpe Ratio, developed by Nobel Prize winner William Sharpe some 50 years ago, does precisely this: it compares the return of an investment to that of an alternative and relates the relative return to the risk of the investment, measured by the standard deviation of returns.